pyproject.toml
[project]
name = "spinachlangapi"
version = "0.1.0"
description = "A quantum programming language and compiler API."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Cedrick Gauthier", email = "spinachlang@proton.me" }]

dependencies = [
  "spinachlang==0.2.3",
  "fastapi==0.118.0",
  "strawberry-graphql[fastapi]==0.282.0",
  "python-dotenv==1.1.1",
  "uvicorn>=0.16,<0.35",
]

[project.scripts]
spinachlangapi = "spinachlangapi.cli:main"


[project.optional-dependencies]
dev = ["pylint>=3.2.0", "pytest>=8.3.0"]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"
